requirements.txt
croniter==1.0.8
Django==3.1.7
google-cloud-texttospeech==2.2.0
